pyproject.toml
[tool.poetry]
name = "test-betting-platform"
version = "0.1.0"
description = ""
authors = ["VB"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.7"
python-json-logger = "^3.2.1"
psycopg = "^3.2.4"
sqlalchemy = "^2.0.37"
uvicorn = "^0.34.0"
asyncio = "^3.4.3"
aio-pika = "^9.5.4"
alembic = "^1.14.1"
greenlet = "^3.1.1"
redis = "^5.2.1"
pyjwt = "^2.10.1"
bcrypt = "^4.2.1"
email-validator = "^2.2.0"
dishka = "^1.4.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
ruff = "^0.9.2"
black = "^24.10.0"
mypy = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
files = "line_provider/, bet_maker/, shared/, tests/"
ignore_missing_imports = true
disallow_untyped_calls = true
disallow_untyped_defs = true
explicit_package_bases = true


[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "B904",
    "E402",
]